refactor(runs): extract shared run-artifact helpers into integration/ai/runs/_engine
The three forensic-artifact flows (detect-scripts / detect-skills / readiness)
each carried local copies of `buildRunDirName`, `BODY_PREVIEW_LIMIT`, and a
near-identical `readBodyPreview` helper — three near-clones drift quickly.

New shared module at `src/integration/ai/runs/_engine/run-artifacts.ts` exports:

  - `buildRunDirName()` — lexicographic-sortable timestamp + base36 suffix.
  - `BODY_PREVIEW_LIMIT` — the 800-char cap on inline body previews.
  - `readRunBodyPreview(runDir, { truncatedSuffix? })` — reads
    `<runDir>/body.txt`, trims, truncates with a caller-supplied suffix.
    Confirm leaves pass the multi-line "[…truncated; full body at …]" trailer;
    error hints (readiness) accept the single-line default.

All three flows now import from this module; ~50 lines of copy-paste removed.
Behaviour preserved — typecheck/lint/test green on the affected suites.
